Separate category grid into its own component
